requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
chardet==4.0.0
idna==3.3
multidict==5.2.0
numpy==1.21.2
pandas==1.3.3
pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
ujson==4.2.0
yarl==1.7.0
